pyproject.toml
[project]
name = "aligncav"
version = "1.0.0"
description = "ML Techniques for Fabry-Pérot Cavity Alignment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Haraprasad Nandi"}
]
keywords = [
    "fabry-perot",
    "cavity-alignment",
    "reinforcement-learning",
    "deep-learning",
    "laser-beam",
    "hermite-gaussian",
    "mode-classification",
    "optics",
    "physics"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "albumentations>=1.3.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "pyserial>=3.5",
    "gymnasium>=1.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
gui = [
    "PyQt5>=5.15.0",
]
tensorboard = [
    "tensorboard>=2.13.0",
    "wandb>=0.15.0",
]
all = [
    "aligncav[dev,gui,tensorboard]",
]

[project.urls]
Homepage = "https://github.com/whiteflakes/aligncav"
Documentation = "https://github.com/whiteflakes/aligncav#readme"
Repository = "https://github.com/whiteflakes/aligncav"
Issues = "https://github.com/whiteflakes/aligncav/issues"

[project.scripts]
aligncav-train-classifier = "aligncav.cli.train_classifier:main"
aligncav-train-rl = "aligncav.cli.train_rl:main"
aligncav-predict = "aligncav.cli.predict:main"
aligncav-simulate = "aligncav.cli.simulate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aligncav"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | MS_2024
    | MS_Thesis
    | MS_2024_other
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["MS_2024", "MS_Thesis", "MS_2024_other"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
